+++ game/src/main/java/com/blaj/openmetin/game/application/common/config/ChannelPropertiesConfig.java
@@ -1,6 +1,8 @@
 package com.blaj.openmetin.game.application.common.config;
 
+import org.joou.UByte;
+
 public interface ChannelPropertiesConfig {
 
-  short channelIndex();
+  UByte channelIndex();
 }
