+++ game/src/main/java/com/blaj/openmetin/game/application/features/deletecharacter/DeleteCharacterCommand.java
@@ -1,6 +1,7 @@
 package com.blaj.openmetin.game.application.features.deletecharacter;
 
 import com.blaj.openmetin.shared.infrastructure.cqrs.Command;
+import org.joou.UByte;
 
-public record DeleteCharacterCommand(short slot, String deleteCode, long sessionId)
+public record DeleteCharacterCommand(UByte slot, String deleteCode, long sessionId)
     implements Command<Void> {}
