+++ game/src/main/java/com/blaj/openmetin/game/application/features/selectcharacter/SelectCharacterCommand.java
@@ -1,5 +1,6 @@
 package com.blaj.openmetin.game.application.features.selectcharacter;
 
 import com.blaj.openmetin.shared.infrastructure.cqrs.Command;
+import org.joou.UByte;
 
-public record SelectCharacterCommand(short slot, long sessionId) implements Command<Void> {}
+public record SelectCharacterCommand(UByte slot, long sessionId) implements Command<Void> {}
