+++ shared/src/main/java/com/blaj/openmetin/shared/application/common/config/TcpConfig.java
@@ -1,12 +1,13 @@
 package com.blaj.openmetin.shared.application.common.config;
 
 import java.net.InetAddress;
+import org.joou.UShort;
 
 public interface TcpConfig {
 
   InetAddress host();
 
-  int port();
+  UShort port();
 
-  int idleSeconds();
+  UShort idleSeconds();
 }
