+++ shared/src/main/java/com/blaj/openmetin/shared/infrastructure/network/server/NettyServerService.java
@@ -44,7 +44,7 @@ public class NettyServerService {
 
     serverChannel =
         serverBootstrap
-            .bind(tcpProperties.host(), tcpProperties.port())
+            .bind(tcpProperties.host(), tcpProperties.port().intValue())
             .syncUninterruptibly()
             .channel();
